code/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
pillow